Project with Metrics & Phoenix API
This source code folder contains version 1 of the Test Flow Application
with major changes on javascript & html files to include metrics
functionality for Version 2.0.

A new folder with the Phoenix Mock API files was also added and may only
be found on this version's repository (you may also access it on the
virtual machine's)

Note that none of this changes interfere with the interconnectivity of
TFS and the APP, meaning that on this aspect it is exactly the same as
version 1, nothing was changed outside the metrics
methods/functionality.

Some libraries were added as well.

Call of metrics is found on:
- \AngularViews\Composites\case_view_control.html

Automation and getMetrics methods may be found on:
-\Controllers\Workspace\workspace.js

There is one issue that emerged with a crossDomain conflict that I
easily fixed locally by opening chrome disabling-web-security, this
issue was what took me the longest time to solve (basically the app url
must be granted permission to access the phoenix api's execution url).
I will be looking for a better solution for this problem not to occur on
the official website without changing chrome's security options.